+++ source/buttons/changeclubseats.js
@@ -10,7 +10,8 @@ module.exports = new ButtonWrapper(mainId, 3000,
 	/** Set the max members and isRecruting for the club with provided id */
 	(interaction, [clubId]) => {
 		const club = getClubDictionary()[clubId];
-		const modal = new ModalBuilder().setCustomId(interaction.id)
+		const modalCustomId = `${SKIP_INTERACTION_HANDLING}${SAFE_DELIMITER}${interaction.id}`;
+		const modal = new ModalBuilder().setCustomId(modalCustomId)
 			.setTitle("Club Membership Settings")
 			.addComponents(
 				new ActionRowBuilder().addComponents(
@@ -23,7 +24,7 @@ module.exports = new ButtonWrapper(mainId, 3000,
 				),
 			);
 		interaction.showModal(modal);
-		interaction.awaitModalSubmit({ filter: submission => submission.customId === interaction.id, time: timeConversion(5, "m", "ms") }).then(modalSubmission => {
+		interaction.awaitModalSubmit({ filter: submission => submission.customId === modalCustomId, time: timeConversion(5, "m", "ms") }).then(modalSubmission => {
 			const { fields } = modalSubmission;
 			const errors = {};