+++ source/commands/_commandDictionary.js
@@ -23,6 +23,7 @@ const commandFiles = [
 	"petition.js",
 	"post-reference.js",
 	"press-kit.js",
+	"proxy-thread",
 	"roles-rundown.js",
 	"roll.js",
 	"server-rules.js",