+++ source/constants.js
@@ -27,10 +27,12 @@ module.exports = {
 	guildId,
 	testGuildId,
 	topicCategoryId,
+	pluralKitId: "466378653216014359",
 	commandIds: {},
 
 	// Internal Convention
 	SAFE_DELIMITER: "→",
+	SKIP_INTERACTION_HANDLING: "❌",
 	imaginaryHorizonsIconURL: "https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png",
 	voiceChannelOptions: {
 		"private": (guild, modRoleId, host) => ({